+++ packages/opencensus-web-all/src/entrypoints/export-initial-load.ts
@@ -14,6 +14,6 @@
  * limitations under the License.
  */
 
-import { exportRootSpanAfterLoadEvent } from '../export-initial-load';
+import { exportRootSpanAfterLoadEvent } from '../index';
 
 exportRootSpanAfterLoadEvent();
