+++ packages/opencensus-web-initial-load/test/index.ts
@@ -16,4 +16,3 @@
 
 import './test-export-initial-load';
 import './test-initial-load-context';
-import './test-initial-load-sampling';
